Rename schema browser file
